# server/swagger/openapi.yaml
...
openapi: 3.0.0
servers:
- url: http://localhost:5000
info:
  description: 'This is a sample E-Commerce 2.0 app. This project requires that you build a fully-functioning e-commerce application REST API that allows users to perform various CRUD operations such as registering an account, browsing products for sale, etc.'
  version: 2.0.0
  title: E-Commerce 2.0 REST API application
  termsOfService: https://www.codecademy.com/about
  contact:
    email: easoftware@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/carts":
    post:
      tags:
      - carts
      summary: Create cart
      description: 'If non-exiatant, done by the app when the user selects a product to purchare.'
      operationId: createCart
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '400':
          description: user_uuid already used, required value missing
        '404':
          description: cart not inserted
        '409':
          description: user_uuid violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartsPost"
    get:
      tags:
      - carts
      summary: Gets all carts
      description: 'Only done by admin'
      operationId: getAllCarts
      responses:
        '200':
          description: successful opertion (No cart rows is OK, array is empty)
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Carts"
  "/carts/{uuid}":
    put:
      tags:
      - carts
      summary: Update cart by cart uuid
      description: ''
      operationId: updateCartByUuid
      parameters:
      - name: uuid
        in: path
        description: 'The uuid of the cart that needs to be updated.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '404':
          description: Invalid Cart Uuid supplied, Cart not found
      requestBody:
        "$ref": "#/components/requestBodies/CartsPut"
    delete:
      tags:
      - carts
      summary: Delete cart
      description: This can only be done by the Admin.
      operationId: deleteCart
      parameters:
      - name: uuid
        in: path
        description: 'The cart id that needs to be deleted.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deleted cart - cart's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '400': 
          description: unknown error
        '404':
          description: Cart not found
        '409':
          description: Cannot delete - constraint error
  "/carts/{uuid}/allitems":          
    delete:
      tags:
      - cart items
      summary: Delete all items in a cart
      description: 'Used in the checkout process, when the user has paid for the order, moving cart items to order items.'
      operationId: cartDeleteAllItems
      parameters:
      - name: uuid
        in: path
        description: 'The cart uuid that that is checking out.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deleted all items - cart's uuid returned 
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Cannot delete
  "/carts/{uuid}/checkout":
    post:
      tags:
      - carts
      summary: Check out operation for cart
      description: 'Starts whole checkout opertation: Payment, move items from cart and cart_items to orders and order_items.'
      operationId: cartCheckout
      parameters:
      - name: uuid
        in: path
        description: 'The cart uuid that that is checking out.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful opertion 
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Orders"
      requestBody:
        "$ref": "#/components/requestBodies/FullCartsPost"
  "/carts/{uuid}/items":
    post:
      tags:
      - cart items
      summary: Create cart item
      description: 'Done by the app when the user selects a product to purchare.'
      operationId: createCartItem
      parameters:
      - name: uuid
        in: path
        description: 'The uuid of the cart, the cart item is being added to.'
        required: true
        schema:
          type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '400':
          description: user_uuid already used, required value missing
        '404':
          description: cart not inserted
        '409':
          description: user_uuid violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPost"
    get:
      tags:
      - cart items
      summary: Gets all items for a cart
      description: 'Done by app when user requests cart'
      operationId: getCartItems
      parameters:
      - name: uuid
        in: path
        description: 'The uuid of the cart.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful opertion 
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Carts"
  "/carts/cart/{uuid}":
    get:
      tags:
      - carts
      summary: Get cart by cart uuid
      description: 'Gets the cart by cart uuid'
      operationId: getCartByCartUuid
      parameters:
      - name: uuid
        in: path
        description: 'The cart uuid that needs to be fetched.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '404':
          description: Invalid Cart Uuid supplied, Cart not found
  "/carts/items/{itemUuid}":
    get:
      tags:
      - cart items
      summary: Get cart items by cart item uuid
      description: 'Gets the cart item by cart item uuid'
      operationId: getCartItemByCartItemUuid
      parameters:
      - name: itemUuid
        in: path
        description: 'The cart item uuid that needs to be fetched.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '404':
          description: Invalid Cart Item Uuid supplied, Cart Item not found
    put:
      tags:
      - cart items
      summary: Update cart item by cart item uuid
      description: 'Updates quantity and modified coumns in cart item row, by cart item id'
      operationId: updateCartItemByUuid
      parameters:
      - name: itemUuid
        in: path
        description: 'The uuid of the cart items that needs to be updated.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Cart_items"
        '404':
          description: Invalid Cart Uuid supplied, Cart not found
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPut"
    delete:
      tags:
      - cart items
      summary: Delete cart item
      description: 'Delete cart item when user makes quantity 0'
      operationId: deleteCartItem
      parameters:
      - name: itemUuid
        in: path
        description: 'The cart item uuid that needs to be deleted.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deleted cart item - cart item's uuid returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Cart Item not found
        '409':
          description: Cannot delete - constraint error
  "/carts/user/{uuid}":
    get:
      tags:
      - carts
      summary: Get cart by user uuid
      description: 'Gets the cart by user uuid'
      operationId: getCartByUserUuid
      parameters:
      - name: uuid
        in: path
        description: 'The user uuid oft he cart that needs to be fetched.'
        required: true
        schema:
          type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '404':
          description: Invalid Cart Uuid supplied, Cart not found
  "/carts/fullcartrow":
    post:
      tags:
      - carts
      summary: Posts a cart row
      description: 'Posts a row to the carts table. Only used in testing'
      operationId: postFullCartRow
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Carts"
        '400':
          description: user_uuid already used, required value missing
        '404':
          description: cart not inserted
        '409':
          description: user_uuid violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/FullCartsPost"
...
components:
  schemas:
    Carts:
      type: object
      properties:
        uuid:
          type: string
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        used_uuid:
          type: string
    Cart_items:
      type: object
      properties:
        uuid:
          type: string
        cart_uuid:
          type: string
        product_uuid:
          type: string
        quantity:
          type: integer
          format: int32
...
    Orders:
      type: object
      properties:
        uuid:
          type: string
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        status:
          type: string
        total_price:
          type: number
          format: Decimal
        user_uuid:
          type: string
...
  requestBodies:
    CartItemsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              product_uuid:
                type: string
              quantity:
                type: integer
                format: int32
    CartItemsPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              quantity:
                type: integer
                format: int32
...
    CartsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              created:
                type: string
                format: date-time
              used_uuid:
                type: string
    CartsPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              quantity:
                type: string
                format: date-time
    FullCartsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              uuid: 
                type: string
              created:
                type: string
                format: date-time
              modified:
                type: string
                format: date-time
              used_uuid:
                type: string
...